restructure debugger util files
